Expose socket connect method that takes query parameters.
In order to connect to a socket.io server requiring an authentication
token to be sent, we needed the ability to send a query string when
connecting.

This commit simply exposes the method on the underling sio_client that
takes query parameters in the form of a string/string map.